+++ access_management/models/project.py
@@ -24,11 +24,14 @@ class ProjectProject(models.Model):
     def _dev_read_only(self) -> bool:
         return (not self.env.su) and self.env.user.has_group("access_management.group_dev")
 
+    def _support_n1_read_only(self) -> bool:
+        return (not self.env.su) and self.env.user.has_group("access_management.group_support_n1")
+
     @api.model
     def get_view(self, view_id=None, view_type="form", **options):
         result = super().get_view(view_id=view_id, view_type=view_type, **options)
 
-        if (self._dg_read_all_active() or self._dev_read_only()) and view_type in {"list", "kanban", "form"}:
+        if (self._dg_read_all_active() or self._dev_read_only() or self._support_n1_read_only()) and view_type in {"list", "kanban", "form"}:
             node = etree.fromstring(result["arch"])
             node.set("create", "0")
             node.set("edit", "0")
@@ -39,12 +42,12 @@ class ProjectProject(models.Model):
 
     @api.model_create_multi
     def create(self, vals_list):
-        if self._dg_read_all_active() or self._dev_read_only():
+        if self._dg_read_all_active() or self._dev_read_only() or self._support_n1_read_only():
             raise AccessError(_("Vous n'êtes pas autorisé à créer des projets."))
         return super().create(vals_list)
 
     def write(self, vals):
-        if self._dg_read_all_active() or self._dev_read_only():
+        if self._dg_read_all_active() or self._dev_read_only() or self._support_n1_read_only():
             raise AccessError(_("Vous n'êtes pas autorisé à modifier des projets."))
         return super().write(vals)
 
@@ -55,8 +58,8 @@ class ProjectProject(models.Model):
             raise AccessError(_("Le groupe PMO n'est pas autorisé à supprimer des projets."))
         if self._df_manager_no_delete():
             raise AccessError(_("Le groupe DF MANAGER n'est pas autorisé à supprimer des projets."))
-        if self._dev_read_only():
-            raise AccessError(_("Le groupe DEV n'est pas autorisé à supprimer des projets."))
+        if self._dev_read_only() or self._support_n1_read_only():
+            raise AccessError(_("Vous n'êtes pas autorisé à supprimer des projets."))
         return super().unlink()
 
 
@@ -78,6 +81,9 @@ class ProjectTask(models.Model):
     def _dev_no_create_delete(self) -> bool:
         return (not self.env.su) and self.env.user.has_group("access_management.group_dev")
 
+    def _support_n1_no_create_delete(self) -> bool:
+        return (not self.env.su) and self.env.user.has_group("access_management.group_support_n1")
+
     @api.model
     def get_view(self, view_id=None, view_type="form", **options):
         result = super().get_view(view_id=view_id, view_type=view_type, **options)
@@ -88,7 +94,7 @@ class ProjectTask(models.Model):
             node.set("edit", "0")
             node.set("delete", "0")
             result["arch"] = etree.tostring(node, encoding="unicode")
-        elif self._dev_no_create_delete() and view_type in {"list", "kanban", "form"}:
+        elif (self._dev_no_create_delete() or self._support_n1_no_create_delete()) and view_type in {"list", "kanban", "form"}:
             node = etree.fromstring(result["arch"])
             node.set("create", "0")
             node.set("delete", "0")
@@ -98,7 +104,7 @@ class ProjectTask(models.Model):
 
     @api.model_create_multi
     def create(self, vals_list):
-        if self._dg_read_all_active() or self._dev_no_create_delete():
+        if self._dg_read_all_active() or self._dev_no_create_delete() or self._support_n1_no_create_delete():
             raise AccessError(_("Vous n'êtes pas autorisé à créer des tâches."))
         return super().create(vals_list)
 
@@ -114,6 +120,6 @@ class ProjectTask(models.Model):
             raise AccessError(_("Le groupe PMO n'est pas autorisé à supprimer des tâches."))
         if self._df_manager_no_delete():
             raise AccessError(_("Le groupe DF MANAGER n'est pas autorisé à supprimer des tâches."))
-        if self._dev_no_create_delete():
-            raise AccessError(_("Le groupe DEV n'est pas autorisé à supprimer des tâches."))
+        if self._dev_no_create_delete() or self._support_n1_no_create_delete():
+            raise AccessError(_("Vous n'êtes pas autorisé à supprimer des tâches."))
         return super().unlink()
